the course notification is on MainActivity now
